+++ packages/react/src/themes/teams/components/Checkbox/checkboxStyles.ts
@@ -1,0 +1,15 @@
+import { ComponentSlotStylesInput, ICSSInJSStyle } from '../../../types'
+import { CheckboxProps, CheckboxState } from '../../../../components/Checkbox/Checkbox'
+import { CheckboxVariables } from './checkboxVariables'
+import getBorderFocusStyles from '../../getBorderFocusStyles'
+
+const checkboxStyles: ComponentSlotStylesInput<CheckboxProps & CheckboxState, CheckboxVariables> = {
+  root: ({ props: p, variables: v, theme: t }): ICSSInJSStyle => ({
+    ...getBorderFocusStyles({
+      siteVariables: t.siteVariables,
+      isFromKeyboard: p.isFromKeyboard,
+    }),
+  }),
+}
+
+export default checkboxStyles