+++ packages/react/test/specs/components/Checkbox/Checkbox-test.tsx
@@ -1,0 +1,6 @@
+import Checkbox from 'src/components/Checkbox/Checkbox'
+import { isConformant } from 'test/specs/commonTests'
+
+describe('Checkbox', () => {
+  isConformant(Checkbox)
+})